+++ dtoolcore/storagebroker.py
@@ -187,6 +187,13 @@ class DiskStorageBroker(object):
             raise(StorageBrokerOSError(
                 "Path already exists: {}".format(self._abspath)
             ))
+
+        # Make sure the parent directory exists.
+        parent, _ = os.path.split(self._abspath)
+        if not os.path.isdir(parent):
+            raise(StorageBrokerOSError(
+                "No such directory: {}".format(parent)))
+
         os.mkdir(self._abspath)
 
         # Create more essential subdirectories.